Updated project structure
Added an include directory for each of the components of the OS, one for the bootloader and one for the kernel.
Also modified the Makefile in order to use "local absolute paths", an "absolute path" to the current project folder.
Lastly added the panic function to handle kernel panics and the sprinf function to print formatted to string buffers